refactor(resolver): split resolve.rs into submodules
Extract three submodules from the 2,355-line resolve.rs:

- resolve/fetch.rs   — fetch_one_packument (the ensure_fetch! spawn
                       body, lifted into a free async fn)
- resolve/vulnerable.rs — is_vulnerable + prefer_non_vulnerable_pick
- resolve/seed.rs    — seed_direct_deps

Also replace the prefetchable! macro with a Resolver method
(is_prefetchable) and pull the post-loop graph assembly into
finalize_resolved_graph. The BFS hot loop and its scope-capturing
macros (ensure_fetch!, note_root_done!) stay inline — extracting
them further would trade ~2k lines of clear linear control flow for
a 20-field context struct or ControlFlow enums.

resolve.rs drops from 2,355 to 2,135 lines. No behavior change;
cargo test suite (~1,700 unit + integration tests) all pass.
